extract models into netstandard library
