+++ DCCollections.Gui/DCCollections.Gui.csproj
@@ -10,6 +10,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\DCCollectionsRequest\DCCollectionsRequest.csproj" />
+    <ProjectReference Include="..\DCCollections.Models\DCCollections.Models.csproj" />
     <PackageReference Include="Microsoft.Extensions.Configuration.Json" Version="8.0.0" />
   </ItemGroup>
