+++ DCCollectionsRequest/CollectionService.cs
@@ -4,6 +4,7 @@ using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
+using RMCollectionProcessor.Models;
 
 namespace RMCollectionProcessor
 {
