+++ DCCollectionsRequest/DCCollectionsRequest.sln
@@ -7,6 +7,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DCCollectionsRequest", "DCC
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DCCollections.Gui", "..\DCCollections.Gui\DCCollections.Gui.csproj", "{A1F6CD65-5B82-4D39-9D9D-6335CB39E2CB}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DCCollections.Models", "..\DCCollections.Models\DCCollections.Models.csproj", "{CE7DB63F-ED4D-4C9A-9CE9-5F3A62DFA22B}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -20,8 +22,12 @@ Global
 		{A1F6CD65-5B82-4D39-9D9D-6335CB39E2CB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{A1F6CD65-5B82-4D39-9D9D-6335CB39E2CB}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A1F6CD65-5B82-4D39-9D9D-6335CB39E2CB}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{A1F6CD65-5B82-4D39-9D9D-6335CB39E2CB}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
+               {A1F6CD65-5B82-4D39-9D9D-6335CB39E2CB}.Release|Any CPU.Build.0 = Release|Any CPU
+                {CE7DB63F-ED4D-4C9A-9CE9-5F3A62DFA22B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+                {CE7DB63F-ED4D-4C9A-9CE9-5F3A62DFA22B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+                {CE7DB63F-ED4D-4C9A-9CE9-5F3A62DFA22B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+                {CE7DB63F-ED4D-4C9A-9CE9-5F3A62DFA22B}.Release|Any CPU.Build.0 = Release|Any CPU
+        EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
