+++ DCCollectionsRequest/DatabaseService.cs
@@ -4,20 +4,8 @@ using System.Data;
 using Microsoft.Data.SqlClient;
 using Microsoft.Extensions.Configuration;
 using System.IO;
+using RMCollectionProcessor.Models;
 
-public class CreditorDefaults
-{
-    public string UserCode { get; set; } = "07675";
-    public string UserName { get; set; } = "AFRICAN UNITY LIFE LIMITED C";
-    public string BankServUserCode { get; set; } = "D457";
-    public string InitiatingParty { get; set; } = "AFRICAN UNITY LIFE LIMITED";
-    public string CreditorName { get; set; } = "AFRICAN UNITY LIFE LIMITED";
-    public string CreditorContactDetails { get; set; } = "+27-0879439612";//087 943 9612
-    public string CreditorAbbreviatedShortName { get; set; } = "AUL";
-    public string CreditorEmail { get; set; } = "MEMBERSHIP@PAMC.CO.ZA";
-    public string CreditorAccountNumber { get; set; } = "0000000004097772529";
-    public string CreditorBankBranch { get; set; } = "632005";
-}
 
 public class DatabaseService
 {
